move fluids to block
