+++ pumpkin/src/block/blocks/fluids/water.rs
@@ -2,20 +2,20 @@ use std::sync::Arc;
 
 use async_trait::async_trait;
 use pumpkin_data::fluid::Fluid;
-use pumpkin_macros::pumpkin_fluid;
+use pumpkin_macros::pumpkin_block;
 use pumpkin_util::math::position::BlockPos;
 
-use crate::{fluid::pumpkin_fluid::PumpkinFluid, world::World};
+use crate::{block::pumpkin_fluid::PumpkinFluid, world::World};
 
 use super::flowing_fluid::FlowingFluid;
 
-#[pumpkin_fluid("minecraft:flowing_lava")]
-pub struct FlowingLava;
+#[pumpkin_block("minecraft:flowing_water")]
+pub struct FlowingWater;
 
-const LAVA_FLOW_SPEED: u16 = 30;
+const WATER_FLOW_SPEED: u16 = 5;
 
 #[async_trait]
-impl PumpkinFluid for FlowingLava {
+impl PumpkinFluid for FlowingWater {
     async fn placed(
         &self,
         world: &World,
@@ -26,7 +26,7 @@ impl PumpkinFluid for FlowingLava {
         _notify: bool,
     ) {
         world
-            .schedule_fluid_tick(fluid.id, *block_pos, LAVA_FLOW_SPEED)
+            .schedule_fluid_tick(fluid.id, *block_pos, WATER_FLOW_SPEED)
             .await;
     }
 
@@ -36,13 +36,13 @@ impl PumpkinFluid for FlowingLava {
 }
 
 #[async_trait]
-impl FlowingFluid for FlowingLava {
+impl FlowingFluid for FlowingWater {
     async fn get_drop_off(&self) -> i32 {
-        2
+        1
     }
 
     async fn get_slope_find_distance(&self) -> i32 {
-        2
+        4
     }
 
     async fn can_convert_to_source(&self, _world: &Arc<World>) -> bool {
