+++ pumpkin/src/main.rs
@@ -59,7 +59,6 @@ pub mod command;
 pub mod data;
 pub mod entity;
 pub mod error;
-pub mod fluid;
 pub mod item;
 pub mod net;
 pub mod plugin;
